CropAI/backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
pillow==10.0.0
numpy==1.24.3
requests==2.31.0